% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_LUSC_demo}
\alias{data_LUSC_demo}
\title{LUSC Demo Data}
\format{
\subsection{\code{data_LUSC_demo}}{

A data frame with 17 rows (genes) and 6 columns (the first column for vertex indices + 5 samples).
}
}
\source{
https://xenabrowser.net/datapages/?dataset=TCGA-LUSC.htseq_fpkm.tsv&host=https\%3A\%2F\%2Fgdc.xenahubs.net&removeHub=https\%3A\%2F\%2Fxena.treehouse.gi.ucsc.edu\%3A443
}
\usage{
data_LUSC_demo
}
\description{
An LUSC normalized gene expression matrix with selected genes.
}
\details{
The demo dataset gives the expression levels of genes which are involved in GO:0006306 (DNA methylation) and GO:0051781 (positive regulation of cell division) and have at least one neighbor on \link{network_demo}.
Each row corresponds to a gene. Each column corresponds to a sample. The first column are the numeric vertex indices.\cr\cr
The dataset is the extracted LUSC part of an aggregated dataset of LGG, GBM, LUSC and LUAD datasets on TCGA after quantile normalization.\cr
LUSC and LUAD are two groups of lung tumors and are usually studied comparatively (as in this demo study).
}
\keyword{datasets}
