% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network_demo}
\alias{network_demo}
\title{Demo Network}
\format{
\subsection{\code{network_demo}}{

An \link[igraph]{igraph-package} graph object.
}
}
\source{
http://hint.yulab.org/download/HomoSapiens/binary/hq/
}
\usage{
network_demo
}
\description{
A gene regulatory network extracted from a HINT (High-quality INTeractomes) binary compilation of protein-protein interactions.
}
\keyword{datasets}
