% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pirwin.hall}
\alias{pirwin.hall}
\title{Distribution function of the Irwin-Hall distribution}
\usage{
pirwin.hall(x, n)
}
\arguments{
\item{x}{A number between 0 and n.}

\item{n}{The number of i.i.d. U(0,1) to sum.}
}
\value{
The CDF of the Irwin-Hall(n) at x.
}
\description{
Computes the distribution function of Irwin-Hall with parameter n.
}
\examples{
pirwin.hall(1, 2)
pirwin.hall(2.5, 4)
}
\references{
Irwin-Hall Distribution. (n.d.). Randomservices.org. Retrieved May 8, 2023, from https://www.randomservices.org/random/special/IrwinHall.html
}
